rating variable type updated

+++ code/MapApp/app/src/main/java/presentation/Adapters/ListViewAdapter.java
@@ -35,7 +35,7 @@ public class ListViewAdapter extends BaseAdapter {
 
        title.setText(data[position][0]);
        description.setText(data[position][1]);
-       float rating = (Integer.valueOf(data[position][2]))/2;
+       float rating = (Float.valueOf(data[position][2]))/2;
        rate_number.setText(Float.toString(rating));
 
        ratingBar.setProgress(Integer.valueOf(data[position][2]));
